restored ui lambda
